app/requirements.txt
Django==3.0
djangorestframework==3.11.1
psycopg2-binary==2.8.5
gunicorn==20.0.4
dj-database-url==0.5.0
django-oscar-api
pysolr